[CHORE][#4]: Configurar módulo core UI
