+++ app/src/main/java/com/dev/ninhohub/MainActivity.kt
@@ -11,7 +11,7 @@ import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
-import com.dev.ninhohub.ui.theme.NinhoHubTheme
+import com.dev.ninhohub.core.ui.theme.NinhoHubTheme
 
 class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
