+++ core-ui/src/main/java/com/dev/ninhohub/core/ui/theme/Color.kt
@@ -1,0 +1,15 @@
+package com.dev.ninhohub.core.ui.theme
+
+import androidx.compose.ui.graphics.Color
+
+val Purple80 = Color(0xFFD0BCFF)
+val PurpleGrey80 = Color(0xFFCCC2DC)
+val Pink80 = Color(0xFFEFB8C8)
+val Purple40 = Color(0xFF6650a4)
+val PurpleGrey40 = Color(0xFF625b71)
+val Pink40 = Color(0xFF7D5260)
+val Blue = Color(0xFF0052CC)
+val DarkBlue = Color(0xFF003366)
+val Green = Color(0xFF008F39)
+val Black = Color(0xFF000000)
+val White = Color(0xFFFFFFFF)
