+++ core-ui/src/main/java/com/dev/ninhohub/core/ui/theme/Theme.kt
@@ -1,6 +1,5 @@
-package com.dev.ninhohub.ui.theme
+package com.dev.ninhohub.core.ui.theme
 
-import android.app.Activity
 import android.os.Build
 import androidx.compose.foundation.isSystemInDarkTheme
 import androidx.compose.material3.MaterialTheme
@@ -9,18 +8,23 @@ import androidx.compose.material3.dynamicDarkColorScheme
 import androidx.compose.material3.dynamicLightColorScheme
 import androidx.compose.material3.lightColorScheme
 import androidx.compose.runtime.Composable
+import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.platform.LocalContext
 
 private val DarkColorScheme = darkColorScheme(
-    primary = Purple80,
+    primary = Blue,
     secondary = PurpleGrey80,
-    tertiary = Pink80
+    tertiary = Pink80,
+    surface = Color(0xFFFFFBFE),
 )
 
 private val LightColorScheme = lightColorScheme(
-    primary = Purple40,
-    secondary = PurpleGrey40,
-    tertiary = Pink40
+    primary = Blue,
+    onPrimary = Color.White,
+    secondary = Green,
+    tertiary = Black,
+    background = Color.White,
+    surface = Color.White
 
     /* Other default colors to override
     background = Color(0xFFFFFBFE),
