+++ core-ui/src/main/java/com/dev/ninhohub/core/ui/theme/Type.kt
@@ -1,4 +1,4 @@
-package com.dev.ninhohub.ui.theme
+package com.dev.ninhohub.core.ui.theme
 
 import androidx.compose.material3.Typography
 import androidx.compose.ui.text.TextStyle
