+++ settings.gradle.kts
@@ -21,4 +21,4 @@ dependencyResolutionManagement {
 
 rootProject.name = "NinhoHub"
 include(":app")
- 
+include(":core-ui")
